New upadtae

+++ app/Product/views.py
@@ -2,12 +2,14 @@ from django.shortcuts import render, redirect, HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth import login, logout, authenticate
-#from Profile.models import Profile
-from .models import Item
+from app.Profiles.models import Profile
+from .models import Item, Category, Brand, SubCategory
 from django import views
-
+from app.Order.models import Order_details
 from .forms import (
+    #LoginForm,
     AddProductForm,
+    EditPrdoductForm,
 )
 # Create your views here.
 
@@ -15,24 +17,69 @@ class ProductListView(LoginRequiredMixin,views.View):
     login_url="/profile/login/"
 
     def get(self,request,*args,**keywrds):
-        if request.user.Profile.Profile_Type != 'SL':
-            return HttpResponse(
-                #status=400,
-                """The user is not--> Seller,
-                can't see the product list"""
-                
-
-            )
-            
-
+        # if request.user.Profile.Profile_Type != 'SL':
+        #     return HttpResponse(
+        #         #status=400,
+        #         "The user is not--> Seller"
+        #     )   
         context={'pl':Item.objects.all()}
         return render(request,'lop.html',context)
+
+
+ 
                                 # 'pl',Product.Item.objects.all()
-        # try:
-        # except Exception as e:
-        #     return HttpResponse('page not found')
+
+def ProductList( request, catid = None):
+    if catid is not None:
+        context={'pl':Item.objects.filter( Category__Sub_Category__id = catid )}
+        return render(request,'p_list.html',context)
+
+    else:
+        context={'pl':Item.objects.all()}
+        return render(request,'p_list.html',context)
 
 
+class CategoryListView(views.View):
+    login_url="/profile/login/"
+
+    def get(self,request, *args, **keywrgs):
+
+       
+   
+    
+        return render(
+                        request,
+                        'catg.html',
+                        context = {
+                            'cl':Category.objects.all()#(Category_Name__startswith=''),
+                            
+                        }
+            )
+class SubCategoryListView(views.View):
+    #login_url="/profile/login/"
+
+    def get(self,request, catid,*args, **keywrgs):
+
+            return render(
+                        request,
+                        'sub_cat.html',
+                        context = {
+                            'cl':Category.objects.get(pk = catid).Sub_Category.all() #(Category_Name__startswith=''), 
+                        }
+            )
+class BrandListView(LoginRequiredMixin, views.View):
+    login_url="/profile/login/"
+
+    def get(self,request, *args, **keywrgs):
+
+        return render(
+                    request,
+                    'bdl.html',
+                    context = {
+                        'bd':Brand.objects.filter(),
+                        
+                    }
+        )
 
 class AddProductView(LoginRequiredMixin,views.View):
     login_url="/profile/login/"
@@ -43,16 +90,18 @@ class AddProductView(LoginRequiredMixin,views.View):
             return HttpResponse(
                 #status=400,
                 "The user is not--> Seller"
-
-            )
-            
+            )   
         if not pid:
             return render(
                             request,
-                            'demo.html',
+                            'addprod.html',
                             context={
                                 'form':AddProductForm(
-                                 #instance= request.user.Item.seller_Product
+                                    #instance=Item.objects.get( id = pid ),
+                                    initial = {
+                                        "Profile" : request.user.Profile,
+                                    }
+                                
                             )
                         }
                     )
@@ -68,8 +117,6 @@ class AddProductView(LoginRequiredMixin,views.View):
                     )  
     def post(self, request,pid=None, *args, **keywrgs):
         
-       
-
         form = AddProductForm(
                 
                 data=request.POST,
@@ -79,7 +126,7 @@ class AddProductView(LoginRequiredMixin,views.View):
             form.save()     #form.create()
             return render(
                     request,
-                    'demo.html',
+                    'addprod.html',
                     context={
                         'form' : form,
                         "errors" : "Profile Updated Sucessfully",
@@ -88,7 +135,7 @@ class AddProductView(LoginRequiredMixin,views.View):
         else:
             return render(
                     request,
-                    'demo.html',
+                    'addprod.html',
                     context={
                         'form' : form,
                     }
@@ -103,7 +150,11 @@ class DeleteProductView(LoginRequiredMixin, views.View):
     def get(self, request, pid=None,*args,**keywrgd):
         # quer=Item.objects.filter(id=pid)
         # quer.delete()
-
+        if request.user.Profile.Profile_Type != 'SL':
+            return HttpResponse(
+                #status=400,
+                "The user is not--> Seller"
+            )   
         Item.objects.get(id=pid).delete()
         
         return redirect('product:list')
@@ -123,9 +174,9 @@ class EditProductView(LoginRequiredMixin, views.View):
             try:
                 return render(
                     request,
-                    'demo.html',
+                    'addprod.html',
                     context={
-                        'form' : AddProductForm(
+                        'form' : EditPrdoductForm(
                             instance=Item.objects.get( id = pid )
                         )
                     }
@@ -156,7 +207,7 @@ class EditProductView(LoginRequiredMixin, views.View):
             )
            
         
-        form = AddProductForm(
+        form = EditPrdoductForm(
             instance=seller_Product,
             data=request.POST,
         )
@@ -164,7 +215,7 @@ class EditProductView(LoginRequiredMixin, views.View):
             form.save()
             return render(
                 request,
-                'demo.html',
+                'addprod.html',
                 context={
                     'form' : form,
                     "errors" : "Profile Updated Sucessfully",
@@ -173,8 +224,36 @@ class EditProductView(LoginRequiredMixin, views.View):
         else:
             return render(
                 request,
-                'demo.html',
+                'addprod.html',
                 context={
                     'form' : form,
                 }
             )
+
+class Pending_order(LoginRequiredMixin,views.View):
+    login_url="/profile/login/"
+
+    def get(self,request,*args,**keywrds):
+        # if request.user.Profile.Profile_Type != 'SL':
+        #     return HttpResponse(
+        #         #status=400,
+        #         "The user is not--> Seller"
+        #     )   
+        context={'order':Order_details.objects.all()}
+        return render(request,'cifodr.html',context)
+
+
+# class Confirm_order(LoginRequiredMixin,views.View):
+#     login_url="/profile/login/"
+
+#     def get(self,request,*args,**keywrds):
+#         if request.user.Profile.Profile_Type != 'SL':
+#             return HttpResponse(
+#                 #status=400,
+#                 "The user is not--> Seller"
+#             )   
+#         context={'cnf':Order_details.objects.filter(Order_type="SC")}
+#         return render(request,'cifodr.html',context)
+
+
+
